+++ packages/appstore/src/components/routes/route-with-sub-routes.jsx
@@ -2,29 +2,12 @@ import React from 'react';
 import { Redirect, Route } from 'react-router-dom';
 import { redirectToLogin, isEmptyObject, routes, removeBranchName, default_title } from '@deriv/shared';
 import { getLanguage } from '@deriv/translations';
-import Page404 from 'Modules/Page404';
 
-const wallet_routes = [
-    routes.wallets,
-    routes.wallets_deposit,
-    routes.wallets_withdrawal,
-    routes.wallets_transfer,
-    routes.wallets_transactions,
-    routes.wallets_compare_accounts,
-    routes.wallets_on_ramp,
-    routes.wallets_reset_balance,
-];
 const RouteWithSubRoutes = route => {
-    const validateRoute = pathname => {
-        if (pathname === '') return true;
-        return route.path === pathname || wallet_routes.some(route => route == pathname);
-    };
-
     const renderFactory = props => {
         let result = null;
 
         const pathname = removeBranchName(location.pathname);
-        const is_valid_route = validateRoute(pathname);
 
         if (route.component instanceof Redirect) {
             let to = route.to;
@@ -50,7 +33,7 @@ const RouteWithSubRoutes = route => {
             result = (
                 <React.Fragment>
                     {has_default_subroute && pathname === route.path && <Redirect to={default_subroute.path} />}
-                    {is_valid_route ? <route.component {...props} routes={route.routes} /> : <Page404 />}
+                    <route.component {...props} routes={route.routes} />
                 </React.Fragment>
             );
         }
