+++ packages/core/src/App/Constants/routes-config.js
@@ -3,12 +3,16 @@ import { Redirect as RouterRedirect } from 'react-router-dom';
 import { makeLazyLoader, routes, moduleLoader } from '@deriv/shared';
 import { Loading } from '@deriv/components';
 import { localize } from '@deriv/translations';
+import RootComponent from '../Components/Routes/root-component';
 import Redirect from 'App/Containers/Redirect';
 import Endpoint from 'Modules/Endpoint';
 
 const CFDCompareAccounts = React.lazy(() =>
     import(/* webpackChunkName: "cfd-compare-accounts" */ '@deriv/cfd/src/Containers/cfd-compare-accounts')
 );
+
+// Error Routes
+const Page404 = React.lazy(() => import(/* webpackChunkName: "404" */ 'Modules/Page404'));
 
 const Trader = React.lazy(() =>
     moduleLoader(() => {
@@ -406,6 +410,13 @@ const getModules = () => {
             is_authenticated: false,
             getTitle: () => localize("Trader's Hub"),
         },
+        {
+            path: routes.traders_hub,
+            component: RootComponent,
+            is_authenticated: false,
+            getTitle: () => localize("Trader's Hub"),
+        },
+        // { component: Page404, getTitle: () => localize('Error 404') },
     ];
 
     return modules;
@@ -419,6 +430,7 @@ const lazyLoadComplaintsPolicy = makeLazyLoader(
 // Order matters
 // TODO: search tag: test-route-parent-info -> Enable test for getting route parent info when there are nested routes
 const initRoutesConfig = () => [
+    { path: routes.index, component: RouterRedirect, getTitle: () => '', to: routes.traders_hub },
     { path: routes.endpoint, component: Endpoint, getTitle: () => 'Endpoint' }, // doesn't need localization as it's for internal use
     { path: routes.redirect, component: Redirect, getTitle: () => localize('Redirect') },
     {
@@ -429,14 +441,17 @@ const initRoutesConfig = () => [
         is_authenticated: true,
     },
     ...getModules(),
-    { path: routes.index, component: RouterRedirect, getTitle: () => '', to: routes.traders_hub },
 ];
 
 let routesConfig;
 
+// // For default page route if page/path is not found, must be kept at the end of routes_config array
+const route_default = { component: Page404, getTitle: () => localize('Error 404') };
+
 const getRoutesConfig = () => {
     if (!routesConfig) {
         routesConfig = initRoutesConfig();
+        routesConfig.push(route_default);
     }
     return routesConfig;
 };
